% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDesc.R
\docType{data}
\name{data.benchmark.calibrator}
\alias{data.benchmark.calibrator}
\title{MiRNA Sequencing Benchmark Calibrator Data}
\format{
A data frame with 10 rows and 54 columns. Here are the examples of the column and row naming rule:
\describe{
  \item{MXF2516_D13}{One sample belonging to MXF and library D with sample ID MXF2516 and library ID D13.}
  \item{cali_11_rc}{Calibrator ID.}
}
}
\usage{
data.benchmark.calibrator
}
\description{
Myxofibrosarcoma (MXF) and pleomorphic malignant fibrous histiocytoma (PMFH) are the two most common and aggressive subtypes of genetically complex soft tissue sarcoma.
This dataset includes the sequencing of 10 calibrators for each sample in the `data.benchmark` data for quality control.
}
\keyword{datasets}
