% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDesc.R
\docType{data}
\name{data.benchmark.pool}
\alias{data.benchmark.pool}
\title{MiRNA Sequencing Benchmark Pooled Data}
\format{
A data frame with 1033 rows and 6 columns. Here are the examples of the column and row naming rule:
\describe{
  \item{MXF_PL1_D16}{One pooled sample belonging to MXF and library D.}
  \item{hsa-let-7a-2*}{Gene ID.}
}
}
\usage{
data.benchmark.pool
}
\description{
Myxofibrosarcoma (MXF) and pleomorphic malignant fibrous histiocytoma (PMFH) are the two most common and aggressive subtypes of genetically complex soft tissue sarcoma.
This dataset includes three libraries used for sequencing 3 pooled tumor samples. Library preparation and read capture were each processed by a single experienced technician in one run.
}
\keyword{datasets}
