% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDesc.R
\docType{data}
\name{data.test.calibrator}
\alias{data.test.calibrator}
\title{MiRNA Sequencing Test Calibrator Data}
\format{
A data frame with 10 rows and 54 columns. Here are the examples of the column and row naming rule:
\describe{
  \item{MXF2516}{One sample belonging to MXF sample ID MXF2516.}
  \item{cali_11_rc}{Calibrator ID.}
}
}
\usage{
data.test.calibrator
}
\description{
Myxofibrosarcoma (MXF) and pleomorphic malignant fibrous histiocytoma (PMFH) are the two most common and aggressive subtypes of genetically complex soft tissue sarcoma.
This dataset includes the sequencing of 10 calibrators for each sample in the `data.test` data for quality control.
}
\keyword{datasets}
